Add files via upload
This is my portfolio using to create HTML and CSS. html is used for structure in this portfolio. Using elements are h1 for heading, p for paragraph and section to declare different parts like a About me,  Skills, Experience, contact etc. to create my structure and CSS is used for giving styles like color alignment and also margin padding etc.. these kinds of elements used to create my portfolio.